refactor(scalars): move input to new UI folder
