+++ packages/design-system/src/connect/components/modal/settings-modal-v2/package-manager/package-manager-input.tsx
@@ -1,5 +1,5 @@
 import { Button } from "#powerhouse";
-import { Input } from "#scalars";
+import { Input } from "#ui";
 import { useCallback, useState } from "react";
 
 export type PackageManagerInputProps = {
