+++ packages/design-system/src/scalars/components/fragments/index.ts
@@ -4,7 +4,6 @@ export * from "./form-description/index.js";
 export * from "./form-group/index.js";
 export * from "./form-label/index.js";
 export * from "./form-message/index.js";
-export * from "./input/index.js";
 export * from "./radio-group-field/index.js";
 export * from "./select-field/index.js";
 export * from "./text-field/index.js";
