+++ packages/design-system/src/ui/components/data-entry/index.ts
@@ -1,0 +1,1 @@
+export { Input, type InputProps } from "./input/input.js";
