+++ packages/design-system/src/ui/components/data-entry/input/input.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../../lib/utils.js";
+import { cn } from "../../../../scalars/lib/utils.js";
 
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
